project pkl

+++ app/Http/Controllers/BukuController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\buku;
 use Illuminate\Http\Request;
+use Alert;
 
 class BukuController extends Controller
 {
@@ -39,19 +40,19 @@ class BukuController extends Controller
     public function store(Request $request)
     {
         $validated = $request->validate([
-            'kode_register' => 'required',
-            'judul' => 'required',
-            'pengarang' => 'required',
-            'penerbit' => 'required',
+            'kode_buku' => 'required',
+            'judul_buku' => 'required',
+            'penulis_buku' => 'required',
+            'penerbit_buku' => 'required',
             'thn_terbit' => 'required',
             'cover' => 'required|image|max:2048',
         ]);
 
         $buku = new Buku;
-        $buku->kode_register = $request->kode_register;
-        $buku->judul = $request->judul;
-        $buku->pengarang = $request->pengarang;
-        $buku->penerbit = $request->penerbit;
+        $buku->kode_buku = $request->kode_buku;
+        $buku->judul_buku = $request->judul_buku;
+        $buku->penulis_buku = $request->penulis_buku;
+        $buku->penerbit_buku = $request->penerbit_buku;
         $buku->thn_terbit = $request->thn_terbit;
         if ($request->hasFile('cover')) {
             $buku->deleteImage();
@@ -101,19 +102,19 @@ class BukuController extends Controller
     public function update(Request $request, $id)
     {
         $validated = $request->validate([
-            'kode_register' => 'required',
-            'judul' => 'required',
-            'pengarang' => 'required',
-            'penerbit' => 'required',
+            'kode_buku' => 'required',
+            'judul_buku' => 'required',
+            'penulis_buku' => 'required',
+            'penerbit_buku' => 'required',
             'thn_terbit' => 'required',
             'cover' => 'required|image|max:2048',
         ]);
 
         $buku = Buku::findOrFail($id);
-        $buku->kode_register = $request->kode_register;
-        $buku->judul = $request->judul;
-        $buku->pengarang = $request->pengarang;
-        $buku->penerbit = $request->penerbit;
+        $buku->kode_buku = $request->kode_buku;
+        $buku->judul_buku = $request->judul_buku;
+        $buku->penulis_buku = $request->penulis_buku;
+        $buku->penerbit_buku = $request->penerbit_buku;
         $buku->thn_terbit = $request->thn_terbit;
         if ($request->hasFile('cover')) {
             $buku->deleteImage();
@@ -123,6 +124,8 @@ class BukuController extends Controller
             $buku->cover = $name;
         }
         $buku->save();
+
+        Alert::success('Good Job', 'Data Edited Successfully');
         return redirect()->route('buku.index');
 
     }
@@ -135,7 +138,14 @@ class BukuController extends Controller
      */
     public function destroy($id)
     {
-        $buku = Buku::findOrFail($id);
+        // $buku = Buku::findOrFail($id);
+
+        if (!Buku::destroy($id)) {
+            return redirect()->back();
+        }
+        Alert::success('Success', 'Data deleted successfully');
+        return redirect()->route('buku.index');
+
         $buku->delete();
         return redirect()->route('buku.index');
